only task left

+++ src/Model/Item.java
@@ -4,19 +4,19 @@ public class Item extends Stockable {
 
     private static int nextId = 1;
 
-    private  int price;
+    private  double price;
     private final String category;
     private  int minStockLevel;
 
 
-    public Item(String name, int quantity, int price, String category, int minStockLevel) {
+    public Item(String name, int quantity, double price, String category, int minStockLevel) {
         super(generateId(), name, quantity);
         this.price = price;
         this.category = category;
         this.minStockLevel = minStockLevel;
     }
 
-    public Item(int id, String name, int quantity,int price, String category, int minStockLevel) {
+    public Item(int id, String name, int quantity,double price, String category, int minStockLevel) {
         super(id, name, quantity);
         syncNextId(id);
         this.price = price;
@@ -24,7 +24,7 @@ public class Item extends Stockable {
         this.minStockLevel = minStockLevel;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
     public String getCategory() {
